requirements.txt
python-dotenv
google-generativeai
openai-whisper
yt-dlp
nltk
moviepy
imageio[ffmpeg]
pydub
